Extract payload load board emission
